extract generateAttributeOperations
